# src/main/resources/schema/openapischema.yaml
...
openapi: 3.1.0
info:
  title: PA Youth Hackathon
  description: Description
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.1
...
servers:
  - url: https://hackathon.swagger.io/api/v3
...
paths:
...
  /user:
    post:
      tags:
        - user
      summary: Create user
      description: Create a new user.
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                  example: 10
                username:
                  type: string
                  example: theUser
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: James
                email:
                  type: string
                  example: john@email.com
                location:
                  type: object
                  properties:
                    longitude:
                      type: string
                    latitude:
                      type: string
                completedStages:
                  type: array
                  description: The stage ids the user have completed
                  items:
                    type: string
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 10
                  username:
                    type: string
                    example: theUser
                  firstName:
                    type: string
                    example: John
                  lastName:
                    type: string
                    example: James
                  email:
                    type: string
                    example: john@email.com
                  location:
                    type: object
                    properties:
                      longitude:
                        type: string
                      latitude:
                        type: string
                  completedStages:
                    type: array
                    description: The stage ids the user have completed
                    items:
                      type: string
  /user/{username}:
    get:
      tags:
        - user
      summary: Get user by user name
      description: ""
      operationId: getUserByName
      parameters:
        - name: username
          in: path
          description: The name that needs to be fetched.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 10
                  username:
                    type: string
                    example: theUser
                  firstName:
                    type: string
                    example: John
                  lastName:
                    type: string
                    example: James
                  email:
                    type: string
                    example: john@email.com
                  location:
                    type: object
                    properties:
                      longitude:
                        type: string
                      latitude:
                        type: string
                  completedStages:
                    type: array
                    description: The stage ids the user have completed
                    items:
                      type: string
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
    put:
      tags:
        - user
      summary: Update user
      description: This can only be done by the logged in user.
      operationId: updateUser
      parameters:
        - name: username
          in: path
          description: name that need to be deleted
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent user in the store
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                  example: 10
                username:
                  type: string
                  example: theUser
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: James
                email:
                  type: string
                  example: john@email.com
                location:
                  type: object
                  properties:
                    longitude:
                      type: string
                    latitude:
                      type: string
                completedStages:
                  type: array
                  description: The stage ids the user have completed
                  items:
                    type: string
      responses:
        default:
          description: successful operation
    delete:
      tags:
        - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
        - name: username
          in: path
          description: The name that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
...